requirements.txt
python-dotenv==1.0.1
fast-mcp==0.0.7